+++ src/menu/debug.ts
@@ -48,7 +48,5 @@ function getMode(): Promise<string> {
 }
 
 export async function run() {
-  const mode = await getMode()
-
-  argon.debug(mode)
+  argon.debug(await getMode())
 }