+++ src/menu/update.ts
@@ -1,3 +1,4 @@
+import * as vscode from "vscode"
 import * as argon from "../argon"
 import { Item } from "."
 
@@ -7,6 +8,37 @@ export const item: Item = {
   action: "update",
 }
 
-export function run() {
-  argon.update()
+function getMode(): Promise<string> {
+  return new Promise((resolve, reject) => {
+    const items = [
+      {
+        label: "$(multiple-windows) Both",
+        mode: "both",
+      },
+      {
+        label: "$(terminal) CLI",
+        mode: "cli",
+      },
+      {
+        label: "$(plug) Plugin",
+        mode: "plugin",
+      },
+    ]
+
+    vscode.window
+      .showQuickPick(items, {
+        title: "Select command mode",
+      })
+      .then((mode) => {
+        if (!mode) {
+          return reject()
+        }
+
+        resolve(mode.mode)
+      })
+  })
+}
+
+export async function run() {
+  argon.update(await getMode())
 }