+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Added
 
 - Completion for optional paths in projects
+- Support for refreshed CLI `update` command
 
 ## [2.0.10] - 2024-08-10